chore(accordion): first pass at updating accordion for sb v7
